1. Add langtorch package in the package path. 2. Refactor OpenAI and OpenAIChat
