+++ build.gradle
@@ -29,27 +29,36 @@ publishing {
 }
 
 dependencies {
+    // Lombok dependency and requires annotationProcessor
     compileOnly 'org.projectlombok:lombok:1.18.26'
     annotationProcessor 'org.projectlombok:lombok:1.18.26'
-}
-
-dependencies {
+    // Logging
+    implementation 'com.google.flogger:flogger:0.7.4'
+    implementation 'com.google.flogger:flogger-system-backend:0.7.4'
+    // Serialization
     implementation 'com.google.code.gson:gson:2.9.1'
     implementation 'org.jetbrains:annotations:24.0.0'
-    implementation 'com.google.acai:acai:1.1'
-    implementation 'com.google.truth:truth:1.1.3'
-    implementation("com.google.guava:guava:31.1-jre")
-    implementation 'com.squareup.okhttp3:okhttp:4.10.0'
-    implementation 'com.theokanning.openai-gpt3-java:service:0.11.1'
-    implementation 'com.google.inject:guice:5.1.0'
-    implementation 'io.github.cdimascio:java-dotenv:5.2.2'
+    // Java Utils
+    implementation "com.google.guava:guava:31.1-jre"
     implementation 'commons-io:commons-io:2.11.0'
+    // Http request
+    implementation 'com.squareup.okhttp3:okhttp:4.10.0'
+    // third-party OpenAI java library
+    implementation 'com.theokanning.openai-gpt3-java:service:0.11.1'
+    // Injection framework
+    implementation 'com.google.inject:guice:5.1.0'
+    // Environment variables
+    implementation 'io.github.cdimascio:java-dotenv:5.2.2'
 }
 
 dependencies {
+    // Mocking
     testImplementation 'org.mockito:mockito-all:1.10.19'
-    testImplementation 'org.junit.jupiter:junit-jupiter-api:5.9.2'
     testRuntimeOnly 'org.junit.jupiter:junit-jupiter-engine:5.9.2'
+    // Testing util
+    implementation 'com.google.acai:acai:1.1'
+    // Assertion
+    implementation 'com.google.truth:truth:1.1.3'
 }
 
 test {
