+++ src/main/java/ai/knowly/langtoch/capability/config/v1/CapabilityConfig.java
@@ -1,0 +1,23 @@
+package ai.knowly.langtoch.capability.config.v1;
+
+import com.google.gson.annotations.SerializedName;
+import lombok.Data;
+
+/** A class for the capability config. */
+@Data
+public class CapabilityConfig {
+  @SerializedName("schema")
+  private int schema;
+
+  @SerializedName("description")
+  private String description;
+
+  @SerializedName("type")
+  private String type;
+
+  @SerializedName("completion")
+  private Completion completion;
+
+  @SerializedName("input")
+  private Input input;
+}
