+++ src/main/java/ai/knowly/langtoch/llm/base/BaseModel.java
@@ -1,0 +1,6 @@
+package ai.knowly.langtoch.llm.base;
+
+/** A model is a model that takes in a prompt and returns a response. */
+public abstract class BaseModel {
+  public abstract String run(String prompt);
+}
