+++ src/main/java/ai/knowly/langtoch/llm/base/chatmodel/BaseChatModel.java
@@ -1,0 +1,14 @@
+package ai.knowly.langtoch.llm.base.chatmodel;
+
+import ai.knowly.langtoch.llm.base.BaseModel;
+import ai.knowly.langtoch.llm.message.BaseChatMessage;
+import java.util.List;
+
+/** A chat model is a model that takes in a list of messages and returns a message. */
+public abstract class BaseChatModel extends BaseModel {
+  // A chat model takes in a list of messages and returns a message.
+  public abstract BaseChatMessage run(List<BaseChatMessage> messages);
+
+  // A chat model takes in a user message and returns a message.
+  public abstract String run(String message);
+}
