+++ src/main/java/ai/knowly/langtoch/llm/message/Role.java
@@ -1,0 +1,8 @@
+package ai.knowly.langtoch.llm.message;
+
+/** A enum for the role of a message. */
+public enum Role {
+  USER,
+  SYSTEM,
+  ASSISTANT
+}
