+++ src/main/java/ai/knowly/langtoch/llm/message/UserMessage.java
@@ -1,0 +1,24 @@
+package ai.knowly.langtoch.llm.message;
+
+import lombok.Builder;
+
+/** A message from the user. */
+@Builder(setterPrefix = "set")
+public final class UserMessage extends BaseChatMessage {
+
+  private final String message;
+
+  public UserMessage(String message) {
+    this.message = message;
+  }
+
+  @Override
+  public String returnMessage() {
+    return message;
+  }
+
+  @Override
+  public Role returnRole() {
+    return Role.USER;
+  }
+}
