+++ src/main/java/ai/knowly/langtoch/tool/Function.java
@@ -1,0 +1,6 @@
+package ai.knowly.langtoch.tool;
+
+/** The common interface for all langtorch functions. */
+public interface Function {
+  Object execute(Object... args);
+}
