--- src/main/java/ai/knowly/llm/base/BaseModel.java
@@ -1,6 +1,0 @@
-package ai.knowly.llm.base;
-
-/** A model is a model that takes in a prompt and returns a response. */
-public abstract class BaseModel {
-  public abstract String run(String prompt);
-}
