--- src/main/java/ai/knowly/llm/message/Role.java
@@ -1,8 +1,0 @@
-package ai.knowly.llm.message;
-
-/** A enum for the role of a message. */
-public enum Role {
-  USER,
-  SYSTEM,
-  ASSISTANT
-}
