--- src/main/java/ai/knowly/tool/Function.java
@@ -1,6 +1,0 @@
-package ai.knowly.tool;
-
-/** The common interface for all langtorch functions. */
-public interface Function {
-  Object execute(Object... args);
-}
