+++ src/test/java/ai/knowly/langtoch/capability/config/example/summarize/prompt.txt
