--- src/test/java/ai/knowly/tool/ToolTest.java
@@ -1,69 +1,0 @@
-package ai.knowly.tool;
-
-import static com.google.common.truth.Truth.assertThat;
-
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.junit.runners.JUnit4;
-
-@RunWith(JUnit4.class)
-public final class ToolTest {
-  @Test
-  public void testTool_sum() {
-    // Arrange.
-    Tool tool =
-        Tool.builder()
-            .setName("Calculator")
-            .setDescription("The tool includes everything related to calculator.")
-            .build();
-
-    Function sum =
-        args -> {
-          int a = (int) args[0];
-          int b = (int) args[1];
-          return a + b;
-        };
-    tool.register("add", sum);
-    // Act.
-    int result = (int) tool.invoke("add", 1, 2);
-    // Assert.
-    assertThat(result).isEqualTo(3);
-  }
-
-  // https://leetcode.com/problems/search-a-2d-matrix/
-  @Test
-  public void testTool_search2DMatrix() {
-    // Arrange.
-    Tool tool =
-        Tool.builder()
-            .setName("LeetcodeSolver")
-            .setDescription("Search a target in a 2D matrix.")
-            .build();
-
-    Function func = args -> searchMatrix((int[][]) args[0], (int) args[1]);
-
-    tool.register("search_2d_matrix", func);
-    // Act.
-    int[][] twoDArray = {
-      {1, 3, 5, 7},
-      {10, 11, 16, 20},
-      {23, 30, 34, 60}
-    };
-    boolean result1 = (boolean) tool.invoke("search_2d_matrix", twoDArray, 3);
-    boolean result2 = (boolean) tool.invoke("search_2d_matrix", twoDArray, 66);
-    // Assert.
-    assertThat(result1).isEqualTo(true);
-    assertThat(result2).isEqualTo(false);
-  }
-
-  public boolean searchMatrix(int[][] matrix, int target) {
-    int i = 0;
-    int j = matrix[0].length - 1;
-    while (i < matrix.length && j >= 0) {
-      if (matrix[i][j] == target) return true;
-      else if (matrix[i][j] > target) j--;
-      else i++;
-    }
-    return false;
-  }
-}
